also search in grid_dir for avail_diags.log
current behavior is to try to load available_diagnostics.log from  'data_dir' 
if the file is not there, use the xmitgcm default available_diagnostics.  

The proposed change would search in the 'grid_dir' for the 
available_diagnostics.log file if it was not found in 'data_dir'.  

In the case when the log file was in neither grid_dir nor data_dir,
we would use the xmitgcm default.