feat: CHECKS ITERATION PARAMETERS WHEN GETTING LLCMODEL DATASETS
When loading datasets using the llcreader the user is able to specify
the iteration start, stop and step. These can also be specified as
attributes in the `BaseLLCModel`. This commit introduces a check that
compares the `iter_start` and `iter_step` fed into `get_dataset` with
the `BaseLLCModel` attributes. If they are found to be incompatible
a warning is triggered.

A warning over an Error was chosen as subsuquent code may still execute
successfully even if there is a mismatch between the user specified and
`BaseLLCModel` derived values for `iter_start` and `iter_stop`.