feat/refac: REORDERS LOADING AND CHECKING OF ITERATION PARAMS
Currently if one tries to load an ECCO portal model by specifying iterations, an error is raised. This is due to the model having attributes `iter_start`, `iter_step` and `iter_stop`. If the user tries to call `get_dataset` with a list `iters` the function
first reads in the `iter_params` from the model's attributes. The function then goes on to check that none of the `iter_params` are defined. Upon finding they have been defined the function raises an error.

The refactor does the following:
(1) Check to see if the user has set any `iter_params` in the function call.
    (i) If they have, check the user hasn't set `iters`. If the user has set iters there's a problem, so raise an error.
    (ii) If no `iters` are set then use the user set `iter_params`

(2) If no user set `iter_params` try reading in a user set `iters` and use this

(3) If still no luck, try reading the model `iter_params`. If any of them are `None` keep going...

(4) Finally try reading the model `iters`.

(5) If nothing has been found then raise an error about the lack of iteration instructions.

(6) Check the `iter_params` or `iters` for consistency with the models attributes.