feat: CHECKS ITERATION PARAMETERS WHEN GETTING LLCMODEL DATASETS (#240)
* feat: CHECKS ITERATION PARAMETERS WHEN GETTING LLCMODEL DATASETS

When loading datasets using the llcreader the user is able to specify
the iteration start, stop and step. These can also be specified as
attributes in the `BaseLLCModel`. This commit introduces a check that
compares the `iter_start` and `iter_step` fed into `get_dataset` with
the `BaseLLCModel` attributes. If they are found to be incompatible
a warning is triggered.

A warning over an Error was chosen as subsuquent code may still execute
successfully even if there is a mismatch between the user specified and
`BaseLLCModel` derived values for `iter_start` and `iter_stop`.

* refactor: CHANGED WARNING TO RUNTIMEWARNING

* test: ADDED TESTS FOR ITERATION CHECKS

* feat: ADDED CHECK ITERS FUNCTION

* feat/refac: REORDERS LOADING AND CHECKING OF ITERATION PARAMS

Currently if one tries to load an ECCO portal model by specifying iterations, an error is raised. This is due to the model having attributes `iter_start`, `iter_step` and `iter_stop`. If the user tries to call `get_dataset` with a list `iters` the function
first reads in the `iter_params` from the model's attributes. The function then goes on to check that none of the `iter_params` are defined. Upon finding they have been defined the function raises an error.

The refactor does the following:
(1) Check to see if the user has set any `iter_params` in the function call.
    (i) If they have, check the user hasn't set `iters`. If the user has set iters there's a problem, so raise an error.
    (ii) If no `iters` are set then use the user set `iter_params`

(2) If no user set `iter_params` try reading in a user set `iters` and use this

(3) If still no luck, try reading the model `iter_params`. If any of them are `None` keep going...

(4) Finally try reading the model `iters`.

(5) If nothing has been found then raise an error about the lack of iteration instructions.

(6) Check the `iter_params` or `iters` for consistency with the models attributes.

* test: UPDATED TESTS

- Refactored the ecco portal tests
- Added an aste portal test

* fix miniconda action

Co-authored-by: Ryan Abernathey <ryan.abernathey@gmail.com>