add code to drop data-variables that are not common to all datasets
This allows diagnostics output to be combined for which some
diagnostics files contain more variables than others. Before this fix,
open_mdsdataset combined these files without a warning but set *all*
variable of files with too few variables to NaN, because of the
behavior of xr.combine_by_coords(...,compat='override',...). This
may be a bug in xr.combine_by_coords.