drop data variables not common to all datasets before combining them (#304)
* add code to drop data-variables that are not common to all datasets

This allows diagnostics output to be combined for which some
diagnostics files contain more variables than others. Before this fix,
open_mdsdataset combined these files without a warning but set *all*
variable of files with too few variables to NaN, because of the
behavior of xr.combine_by_coords(...,compat='override',...). This
may be a bug in xr.combine_by_coords.

* change variable name and reduce output to a summary

* protect against empty sets, add test for dropping uncommon vars

* move test earlier in the list, and it passes ... weird

* delete extra whitespace

Co-authored-by: Timothy Smith <timsmith204@utexas.edu>